Refactor: rename snake_case to CammelCase
