+++ components/Board.tsx
